# backend/go/src/app/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Game Master Notes Backend API",
    "version": "0.1.0",
    "description": "HTTP API for Game Master Notes domain services."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Default local server (runtime metadata is dynamically injected)."
    }
  ],
...
  "paths": {
...
    "/campaigns": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "List campaigns",
        "parameters": [
          { "$ref": "#/components/parameters/Offset" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListCampaignsResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Campaigns"],
        "summary": "Create campaign",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCampaignRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CampaignResponse" }
              }
            }
          }
        }
      }
    },
    "/campaigns/{id}": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "Get campaign by id",
        "parameters": [
          { "$ref": "#/components/parameters/ID" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CampaignResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Campaigns"],
        "summary": "Update campaign",
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCampaignRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CampaignResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Campaigns"],
        "summary": "Delete campaign",
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": { "204": { "description": "No Content" } }
      }
    },
    "/campaigns/{id}/players": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "List players in campaign",
        "parameters": [
          { "$ref": "#/components/parameters/ID" },
          { "$ref": "#/components/parameters/Offset" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListCampaignPlayersResponse" }
              }
            }
          }
        }
      }
    },
    "/campaigns/{id}/players/{player_id}": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "Get campaign-player relation",
        "parameters": [
          { "$ref": "#/components/parameters/ID" },
          { "$ref": "#/components/parameters/PlayerID" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CampaignPlayerResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Campaigns"],
        "summary": "Add player to campaign",
        "parameters": [
          { "$ref": "#/components/parameters/ID" },
          { "$ref": "#/components/parameters/PlayerID" }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CampaignPlayerResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Campaigns"],
        "summary": "Remove player from campaign",
        "parameters": [
          { "$ref": "#/components/parameters/ID" },
          { "$ref": "#/components/parameters/PlayerID" }
        ],
        "responses": { "204": { "description": "No Content" } }
      }
    },
    "/players/{id}/campaigns": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "List campaigns for player",
        "parameters": [
          { "$ref": "#/components/parameters/ID" },
          { "$ref": "#/components/parameters/Offset" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListCampaignPlayersResponse" }
              }
            }
          }
        }
      }
    },
    "/sessions": {
      "get": {
        "tags": ["Sessions"],
        "summary": "List sessions",
        "parameters": [
          { "$ref": "#/components/parameters/Offset" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListSessionsResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Sessions"],
        "summary": "Create session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSessionRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          }
        }
      }
    },
    "/sessions/{id}": {
      "get": {
        "tags": ["Sessions"],
        "summary": "Get session by id",
        "parameters": [
          { "$ref": "#/components/parameters/ID" },
          { "$ref": "#/components/parameters/IncludeDeleted" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Sessions"],
        "summary": "Update session",
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateSessionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Sessions"],
        "summary": "Delete session",
        "parameters": [{ "$ref": "#/components/parameters/ID" }],
        "responses": { "204": { "description": "No Content" } }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Resource identifier (ULID string)."
      },
      "PlayerID": {
        "name": "player_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "Offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "default": 0
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "default": 20
        }
      },
      "IncludeDeleted": {
        "name": "include_deleted",
        "in": "query",
        "required": false,
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
    },
    "schemas": {
...
      "CreateCampaignRequest": {
        "type": "object",
        "properties": {
          "world_id": { "type": "string" },
          "name": { "type": "string" },
          "start_date": { "type": "string", "format": "date", "nullable": true },
          "end_date": { "type": "string", "format": "date", "nullable": true }
        }
      },
      "UpdateCampaignRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "start_date": { "type": "string", "format": "date", "nullable": true },
          "end_date": { "type": "string", "format": "date", "nullable": true }
        }
      },
      "CampaignResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "world_id": { "type": "string" },
          "name": { "type": "string" },
          "start_date": { "type": "string", "format": "date", "nullable": true },
          "end_date": { "type": "string", "format": "date", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true },
          "version": { "type": "integer", "format": "int32" }
        }
      },
      "ListCampaignsResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/CampaignResponse" } },
          "offset": { "type": "integer", "format": "int32" },
          "limit": { "type": "integer", "format": "int32" },
          "include_deleted": { "type": "boolean" }
        }
      },
      "CampaignPlayerResponse": {
        "type": "object",
        "properties": {
          "campaign_id": { "type": "string" },
          "player_id": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ListCampaignPlayersResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/CampaignPlayerResponse" } },
          "offset": { "type": "integer", "format": "int32" },
          "limit": { "type": "integer", "format": "int32" },
          "include_deleted": { "type": "boolean" }
        }
      },
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "campaign_id": { "type": "string" },
          "played_on": { "type": "string", "format": "date", "nullable": true },
          "summary_md": { "type": "string" }
        }
      },
      "UpdateSessionRequest": {
        "type": "object",
        "properties": {
          "played_on": { "type": "string", "format": "date", "nullable": true },
          "summary_md": { "type": "string" }
        }
      },
      "SessionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "campaign_id": { "type": "string" },
          "played_on": { "type": "string", "format": "date", "nullable": true },
          "summary_md": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "deleted_at": { "type": "string", "format": "date-time", "nullable": true },
          "version": { "type": "integer", "format": "int32" }
        }
      },
      "ListSessionsResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/SessionResponse" } },
          "offset": { "type": "integer", "format": "int32" },
          "limit": { "type": "integer", "format": "int32" },
          "include_deleted": { "type": "boolean" }
        }
      },
...
    }
  }
}